+++ CHANGES.md
@@ -8,6 +8,7 @@
 - Add the `dft` field combinator.
 - Add the `json` combinator.
 - Add the `enum` combinator.
+- Add the `satisfies` combinator.
 
 # 2.0.0 (2024-01-12)